requirements.txt
streamlit==1.39.0
pandas==2.2.3
google-auth==2.35.0
gspread==6.1.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
requests==2.32.3
